♻️ refactor: firebase 모듈 분리
