+++ api/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     implementation project(':core:core-infra-redis');
     implementation project(':core:core-infra-email');
     implementation project(':core:core-infra-portone');
+    implementation project(':core:core-infra-firebase');
 
     // Core
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
@@ -32,10 +33,6 @@ dependencies {
 
     // Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
-
-    // FCM
-    implementation 'com.google.firebase:firebase-admin:9.1.1'
-    implementation group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.2.2'
 }
 bootJar { enabled = true }
 jar { enabled = false }
