+++ api/src/main/java/com/sponus/sponusbe/domain/propose/service/ProposeService.java
@@ -14,10 +14,10 @@ import com.sponus.coredomain.domain.propose.Propose;
 import com.sponus.coredomain.domain.propose.ProposeAttachment;
 import com.sponus.coredomain.domain.propose.ProposeStatus;
 import com.sponus.coredomain.domain.propose.repository.ProposeRepository;
+import com.sponus.coreinfrafirebase.FirebaseService;
 import com.sponus.coreinfras3.S3Util;
 import com.sponus.sponusbe.domain.announcement.exception.AnnouncementErrorCode;
 import com.sponus.sponusbe.domain.announcement.exception.AnnouncementException;
-import com.sponus.sponusbe.domain.notification.service.FirebaseService;
 import com.sponus.sponusbe.domain.propose.dto.request.ProposeCreateRequest;
 import com.sponus.sponusbe.domain.propose.dto.request.ProposeStatusUpdateRequest;
 import com.sponus.sponusbe.domain.propose.dto.request.ProposeUpdateRequest;
