+++ api/src/main/java/com/sponus/sponusbe/domain/report/service/ReportService.java
@@ -14,8 +14,8 @@ import com.sponus.coredomain.domain.report.Report;
 import com.sponus.coredomain.domain.report.ReportAttachment;
 import com.sponus.coredomain.domain.report.ReportImage;
 import com.sponus.coredomain.domain.report.repository.ReportRepository;
+import com.sponus.coreinfrafirebase.FirebaseService;
 import com.sponus.coreinfras3.S3Util;
-import com.sponus.sponusbe.domain.notification.service.FirebaseService;
 import com.sponus.sponusbe.domain.propose.exception.ProposeErrorCode;
 import com.sponus.sponusbe.domain.report.dto.request.ReportCreateRequest;
 import com.sponus.sponusbe.domain.report.dto.request.ReportUpdateRequest;
