+++ core/core-infra-firebase/src/main/resources/application-firebase.yml
@@ -1,0 +1,4 @@
+firebase:
+  fcmUrl: ${FIREBASE_URL}
+  firebaseConfigPath: ${FIREBASE_PATH}
+  scope: https://www.googleapis.com/auth/cloud-platform
