+++ settings.gradle
@@ -6,3 +6,4 @@ include 'core:core-infra-s3'
 include 'core:core-infra-redis'
 include 'core:core-infra-email'
 include 'core:core-infra-portone'
+include 'core:core-infra-firebase'
